feat(atlas-2024-odfr-hi): ATLAS HI Open Data for Research
First commit of files for ATLAS HI Open Data for Research. This should
be everything we need for creating the heavy ion open data for research
records. The record IDs and DOIs are place-holders until they are
issued, but everything else should be final. All files are at CERN
already.